Cleared last of the warnings within our project

+++ KentDAW/Source/core/MidiManager.cpp
@@ -49,7 +49,7 @@ int MidiManager::getCurrentNoteNumber()
     return noteNumber;
 }
 
-float MidiManager::getCurrentNoteHz()
+double MidiManager::getCurrentNoteHz()
 {
     return noteHz;
 }
